누적 랜덤 구현
