+++ src/main/java/com/game/gb5/domain/character/Character.java
@@ -1,7 +1,10 @@
-package com.game.gb5.entity.character;
+package com.game.gb5.domain.character;
+
+import javax.persistence.Entity;
 
 import lombok.Data;
 
+@Entity
 @Data
 public class Character {
 	private String name;
