+++ src/main/java/com/game/gb5/domain/character/SpecialAbility.java
@@ -1,0 +1,4 @@
+package com.game.gb5.domain.character;
+
+public class SpecialAbility {
+}
