+++ src/main/java/com/game/gb5/domain/voucher/Voucher.java
@@ -1,0 +1,4 @@
+package com.game.gb5.domain.voucher;
+
+public class Voucher {
+}
