--- src/main/java/com/game/gb5/entity/character/PersonalAbility.java
@@ -1,4 +1,0 @@
-package com.game.gb5.entity.character;
-
-public class PersonalAbility {
-}
