--- src/main/java/com/game/gb5/entity/character/SpecialAbility.java
@@ -1,4 +1,0 @@
-package com.game.gb5.entity.character;
-
-public class SpecialAbility {
-}
