--- src/main/java/com/game/gb5/entity/voucher/Voucher.java
@@ -1,4 +1,0 @@
-package com.game.gb5.entity.voucher;
-
-public class Voucher {
-}
