+++ src/main/java/com/game/gb5/service/CharacterRecruitmentService.java
@@ -2,7 +2,7 @@ package com.game.gb5.service;
 
 import java.util.List;
 
-import com.game.gb5.entity.character.Character;
+import com.game.gb5.domain.character.Character;
 
 public class CharacterRecruitmentService {
 	private final int STARS_COUNT = 250; // 별 소모 개수
